server/ocr_nlp/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
pillow
pytesseract
spacy
openai-whisper
torch
